python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# AI/ML Libraries
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.4

# NLP Libraries (compatible with Python 3.12)
transformers==4.40.0
torch==2.3.0
sentence-transformers==2.7.0

# OCR
pytesseract==0.3.10
Pillow==10.3.0
opencv-python==4.9.0.80

# Google Books API
google-api-python-client==2.108.0

# Google Calendar API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Utilities
httpx==0.27.0
redis==5.0.1
